Add files via upload
Enum in java are the static final constants which are declared as a constant instances of the corresponding Enum class.There are various methods in Enum like values(), valuesof(), ordinal(), we can define custom fields and methods in Enum class and can access them via object or instance created in the main method(). Following code also have a thread class extending thread in which run method of the thread is overrided in which value of Enum  [].length  of i is printed.